--- app/upgrades/v2.1.0/constant.go
@@ -1,3 +1,0 @@
-package v2_1_0 //nolint:revive,stylecheck // app version
-
-const Name = "v2.1.0"
